Benchmark/code/embeddingGeneration/requirements.txt
attrs==19.3.0
audiotools==3.1.1
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
blinker==1.4
blis==0.7.5
Brlapi==0.7.0
catalogue==2.0.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cupy-cuda113==10.2.0
cymem==2.0.6
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
fasteners==0.14.1
fastrlock==0.8
future==0.18.2
GDAL==3.0.4
gdown==5.0.1
hyperlink==19.0.0
idna==2.8
incremental==16.10.1
Jinja2==3.0.3
langcodes==3.3.0
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
MarkupSafe==2.1.0
monotonic==1.5
murmurhash==1.0.6
netifaces==0.10.4
numpy==1.22.2
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pathy==0.6.1
pexpect==4.6.0
Pillow==7.0.0
preshed==3.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
pydantic==1.8.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
screen-resolution-extra==0.0.0
service-identity==18.1.0
six==1.14.0
smart-open==5.2.1
soupsieve==2.5
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
ssh-import-id==5.10
thinc==8.0.13
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.63.0
Twisted==18.9.0
typer==0.4.0
typing-extensions==4.1.1
ufw==0.36
urllib3==1.25.8
urwid==2.0.1
wasabi==0.9.0
xkit==0.0.0
zope.interface==4.7.1
